requirements.txt
Django>=3.0.6
djangorestframework
django-phonenumber-field
phonenumbers
